# Faster computation of cached Frobenius powers